backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib
datetime
python-jose[cryptography]>=3.3.0
mysql-connector-python
email-validator
python-multipart
python-dotenv
alembic